Convert indexprovider internals to hooks
