+++ src/components/IndexProvider/IndexProvider.tsx
@@ -1,96 +1,39 @@
-import React, {useRef, useMemo, useCallback} from 'react';
+import React, {useMemo} from 'react';
 
-import {useI18n} from '../../utilities/i18n';
 import {
   IndexContext,
   IndexRowContext,
   IndexSelectionChangeContext,
-  SELECT_ALL_ITEMS,
   IndexProviderProps,
-  SelectionType,
-  Range,
+  useBulkSelectionData,
+  useHandleBulkSelection,
 } from '../../utilities/index-provider';
-
-type HandleSelectionChange = (
-  selectionType: SelectionType,
-  toggleType: boolean,
-  selection?: string | Range,
-  sortOrder?: number,
-) => void;
 
 export function IndexProvider({
   children,
   resourceName: passedResourceName,
   loading,
-  onSelectionChange = noop,
+  onSelectionChange,
   selectedItemsCount,
   itemCount,
   hasMoreItems,
   condensed,
 }: IndexProviderProps) {
-  const selectable = Boolean(selectedItemsCount);
-
-  const i18n = useI18n();
-  const lastSelected = useRef<number | null>(null);
-
-  const defaultResourceName = {
-    singular: i18n.translate('Polaris.IndexProvider.defaultItemSingular'),
-    plural: i18n.translate('Polaris.IndexProvider.defaultItemPlural'),
-  };
-
-  const selectMode = selectedItemsCount === 'All' || selectedItemsCount > 0;
-
-  const handleSelectionChange: HandleSelectionChange = useCallback(
-    (
-      selectionType: SelectionType,
-      toggleType: boolean,
-      selection?: string | Range,
-      sortOrder?: number,
-    ) => {
-      const prevSelected = lastSelected.current;
-
-      if (SelectionType.Multi && typeof sortOrder === 'number') {
-        lastSelected.current = sortOrder;
-      }
-
-      if (
-        selectionType === SelectionType.Single ||
-        (selectionType === SelectionType.Multi &&
-          (typeof prevSelected !== 'number' || typeof sortOrder !== 'number'))
-      ) {
-        onSelectionChange(SelectionType.Single, toggleType, selection);
-      } else if (selectionType === SelectionType.Multi) {
-        const min = Math.min(prevSelected as number, sortOrder as number);
-        const max = Math.max(prevSelected as number, sortOrder as number);
-        onSelectionChange(selectionType, toggleType, [min, max]);
-      } else if (
-        selectionType === SelectionType.Page ||
-        selectionType === SelectionType.All
-      ) {
-        onSelectionChange(selectionType, toggleType);
-      }
-    },
-    [onSelectionChange],
-  );
-
-  let bulkSelectState: boolean | 'indeterminate' | undefined = 'indeterminate';
-  if (!selectedItemsCount || selectedItemsCount === 0) {
-    bulkSelectState = undefined;
-  } else if (
-    selectedItemsCount === SELECT_ALL_ITEMS ||
-    selectedItemsCount === itemCount
-  ) {
-    bulkSelectState = true;
-  }
-
-  const resourceName = passedResourceName
-    ? passedResourceName
-    : defaultResourceName;
-
-  const paginatedSelectAllText = getPaginatedSelectAllText();
-
-  const bulkActionsLabel = getBulkActionsLabel();
-  const bulkActionsAccessibilityLabel = getBulkActionsAccessibilityLabel();
+  const {
+    paginatedSelectAllText,
+    bulkActionsLabel,
+    bulkActionsAccessibilityLabel,
+    resourceName,
+    selectMode,
+    bulkSelectState,
+    selectable,
+  } = useBulkSelectionData({
+    selectedItemsCount,
+    itemCount,
+    hasMoreItems,
+    resourceName: passedResourceName,
+  });
+  const handleSelectionChange = useHandleBulkSelection({onSelectionChange});
 
   const contextValue = useMemo(
     () => ({
@@ -134,75 +77,10 @@ export function IndexProvider({
   return (
     <IndexContext.Provider value={contextValue}>
       <IndexRowContext.Provider value={rowContextValue}>
-        <IndexSelectionChangeContext.Provider
-          value={handleSelectionChange as any}
-        >
+        <IndexSelectionChangeContext.Provider value={handleSelectionChange}>
           {children}
         </IndexSelectionChangeContext.Provider>
       </IndexRowContext.Provider>
     </IndexContext.Provider>
   );
-
-  function getPaginatedSelectAllText() {
-    if (!selectable || !hasMoreItems) {
-      return;
-    }
-
-    if (selectedItemsCount === SELECT_ALL_ITEMS) {
-      return i18n.translate('Polaris.IndexProvider.allItemsSelected', {
-        itemsLength: itemCount,
-        resourceNamePlural: resourceName.plural.toLocaleLowerCase(),
-      });
-    }
-  }
-
-  function getBulkActionsLabel() {
-    const selectedItemsCountLabel =
-      selectedItemsCount === SELECT_ALL_ITEMS
-        ? `${itemCount}+`
-        : selectedItemsCount;
-
-    return i18n.translate('Polaris.IndexProvider.selected', {
-      selectedItemsCount: selectedItemsCountLabel,
-    });
-  }
-
-  function getBulkActionsAccessibilityLabel() {
-    const totalItemsCount = itemCount;
-    const allSelected = selectedItemsCount === totalItemsCount;
-
-    if (totalItemsCount === 1 && allSelected) {
-      return i18n.translate(
-        'Polaris.IndexProvider.a11yCheckboxDeselectAllSingle',
-        {
-          resourceNameSingular: resourceName.singular,
-        },
-      );
-    } else if (totalItemsCount === 1) {
-      return i18n.translate(
-        'Polaris.IndexProvider.a11yCheckboxSelectAllSingle',
-        {
-          resourceNameSingular: resourceName.singular,
-        },
-      );
-    } else if (allSelected) {
-      return i18n.translate(
-        'Polaris.IndexProvider.a11yCheckboxDeselectAllMultiple',
-        {
-          itemsLength: itemCount,
-          resourceNamePlural: resourceName.plural,
-        },
-      );
-    } else {
-      return i18n.translate(
-        'Polaris.IndexProvider.a11yCheckboxSelectAllMultiple',
-        {
-          itemsLength: itemCount,
-          resourceNamePlural: resourceName.plural,
-        },
-      );
-    }
-  }
 }
-
-function noop() {}
